requirements.txt
Flask==2.0.1
tensorflow==2.5.0
Flask-Dropzone==1.5.4
Pillow==8.2.0